requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
yt-dlp==2024.5.27
APScheduler==3.10.4
Pillow==10.3.0
aiohttp==3.9.5
aiofiles==23.2.1
python-multipart==0.0.9
watchdog==4.0.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
